LS: Refactor `get_node_and_lookup_items` and `find_node_module`
These two functions have been decoupled, documented and moved to the new
 `LsSemanticGroup` extension trait. Then, all usage places have also
 been moderately cleaned up.

commit-id:710e9404
