+++ crates/cairo-lang-language-server/src/ide/semantic_highlighting/mod.rs
@@ -1,5 +1,4 @@
 use cairo_lang_compiler::db::RootDatabase;
-use cairo_lang_filesystem::ids::FileId;
 use cairo_lang_filesystem::span::TextOffset;
 use cairo_lang_parser::db::ParserGroup;
 use cairo_lang_semantic::db::SemanticGroup;
@@ -37,7 +36,7 @@ pub fn semantic_highlight_full(
     };
 
     let mut data: Vec<SemanticToken> = Vec::new();
-    SemanticTokensTraverser::default().find_semantic_tokens(db.upcast(), file, &mut data, node);
+    SemanticTokensTraverser::default().find_semantic_tokens(db.upcast(), &mut data, node);
     Some(SemanticTokensResult::Tokens(SemanticTokens { result_id: None, data }))
 }
 
@@ -54,8 +53,7 @@ struct SemanticTokensTraverser {
 impl SemanticTokensTraverser {
     pub fn find_semantic_tokens(
         &mut self,
-        db: &dyn SemanticGroup,
-        file_id: FileId,
+        db: &(dyn SemanticGroup + 'static),
         data: &mut Vec<SemanticToken>,
         node: SyntaxNode,
     ) {
@@ -72,7 +70,7 @@ impl SemanticTokensTraverser {
                 let maybe_semantic_kind = self
                     .offset_to_kind_lookahead
                     .remove(&node.offset())
-                    .or_else(|| SemanticTokenKind::from_syntax_node(db, file_id, node));
+                    .or_else(|| SemanticTokenKind::from_syntax_node(db, node));
                 if let Some(semantic_kind) = maybe_semantic_kind {
                     let EncodedToken { delta_line, delta_start } = self.encoder.encode(width);
                     data.push(SemanticToken {
@@ -125,7 +123,7 @@ impl SemanticTokensTraverser {
                     _ => {}
                 }
                 for child in children.iter() {
-                    self.find_semantic_tokens(db, file_id, data, child.clone());
+                    self.find_semantic_tokens(db, data, child.clone());
                 }
             }
         }
