+++ crates/cairo-lang-language-server/src/lang/mod.rs
@@ -1,3 +1,4 @@
 pub mod diagnostics;
 pub mod lsp;
+pub mod semantic;
 pub mod syntax;
